Changes most int variable types into float variable types

+++ python-maximum-of-two-numbers.py
@@ -3,8 +3,8 @@
 # in which case it prints the string "Those numbers are equal". 
 # If they aren't equal, it finds the larger number of those two and prints it.
 
-num1 = int(input('First number: '))
-num2 = int(input('Second number: '))
+num1 = float(input('First number: '))
+num2 = float(input('Second number: '))
 
 if num1 == num2:
     print('Those numbers are equal')
